feat(timeline): stabilise governor-limit tooltip ordering and values
- sort tooltip rows by each metric's global peak (not the value at the cursor) so rows hold a fixed slot
  instead of reshuffling as the pointer moves
- always render the (used / limit) value, even at 0% before a metric's first observation, so the
  limit/headroom is always visible; carry the fixed per-series limit on the classified metric
- extract the ApexLog → HeatStripTimeSeries transform out of ApexLogTimeline into apex-limit-series.ts,
  keeping metric-strip/ Apex-agnostic
